project-server/requirements.txt
azure==4.0.0
azure-iot-device==2.0.0rc12
azure-iothub-device-client==1.4.6
pychalk==2.0.1
